gv-notify: Support arbitrary values for `--past`
Handling of the `--past` command-line option expected to only ever
need to look back one month.  However:
* This was not documented.
* This was not enforced (and enforcing it is non-trivial since it
  either would depend on the number of days in the month or depend on
  setting an arbitrary limit).

Fix the code to support an arbitrary amount of time for `--past`
instead. (Setting it to large values is still not advised due to
the large amount of data that would need to be processed, however.)